+++ src/Savegame/BattleUnit.cpp
@@ -1724,11 +1724,11 @@ void BattleUnit::playDeathSound(bool fleshWound) const
 			{
 				default:
 				case GENDER_MALE:
-					soundId = static_cast<int>(ResourcePack::MALE_SCREAM[static_cast<size_t>(RNG::seedless(0,2))]);
+					soundId = ResourcePack::MALE_SCREAM[static_cast<size_t>(RNG::seedless(0,2))];
 					break;
 
 				case GENDER_FEMALE:
-					soundId = static_cast<int>(ResourcePack::FEMALE_SCREAM[static_cast<size_t>(RNG::seedless(0,2))]);
+					soundId = ResourcePack::FEMALE_SCREAM[static_cast<size_t>(RNG::seedless(0,2))];
 			}
 		}
 		else
